fserv v0.19
- http: keep-alive connection didn't work for 304 response
- http: fixed segfault in some cases when there's an error in configuration
- http.static-file: stale cached file descriptors weren't handled properly
- net.connect: windows: stale keep-alive connections weren't handled properly
- net.listen: fixed segfault when log context wasn't reset

+++ src/test/http-test.c
@@ -112,7 +112,7 @@ static int test_http_recvbody(ffskt sk, ffhttp_response *resp, size_t rresp)
 			, sbody.len, &sbody);
 
 		if (resp->h.cont_len != -1) {
-			x(resp->h.cont_len >= sbody.len);
+			x((uint64)resp->h.cont_len >= sbody.len);
 			resp->h.cont_len -= sbody.len;
 			if (resp->h.cont_len == 0)
 				break;
